requirements.txt
streamlit
transformers
Pillow
openai
